Separate file for serializers
